+++ interrupt.h
@@ -1,5 +1,11 @@
 #pragma once
 
+#include <linux/kvm.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <fcntl.h>
+#include <sys/ioctl.h>
+
 struct irr_queue {
     int buff[1000];
     int last;
@@ -7,3 +13,9 @@ struct irr_queue {
 
 void enq_irr(struct irr_queue *irr, int value);
 int deq_irr(struct irr_queue *irr);
+void inject_interrupt(int vcpufd, int irq);
+
+// struct interrupt_buffer {
+//     int head, end, count, max;
+//     int *buff;
+// };
